Improve angle calculation
It's still not great, but good enough for now. Added corrections for any margin of error greater than 0.1% so that the "apparent" velocity will, hopefully, always be the same. For the most part, the ball reflects and/or refracts as you would expect it to, at least when bouncing off the paddle.